Remote extraneous cast.  Go to bed!

+++ Source/VirtualMachine/Values/NumberValue.cpp
@@ -36,7 +36,7 @@ namespace Powder
 			case MathInstruction::MathOp::FACTORIAL:
 			{
 				// TODO: There is actually a factorial for non-integers, and why not support it?  It might not be trivial to calculate, though.
-				double rounded = double(::round(this->number));
+				double rounded = ::round(this->number);
 				double factorial = 1.0;
 				while (rounded-- > 0.0)
 					factorial *= rounded;
